feat(*): myapp -> rr_app, raytracer. Clean code.
