+++ src/scenes/scene_manager.h
