+++ template/common.h
@@ -1,6 +1,6 @@
 // default screen resolution
-#define SCRWIDTH	300
-#define SCRHEIGHT	300
+#define SCRWIDTH	1920
+#define SCRHEIGHT	1080
 #define ASPECT_RATIO SCRWIDTH / SCRHEIGHT
 
 // IMPORTANT NOTE ON OPENCL COMPATIBILITY ON OLDER LAPTOPS:
